sideBar suspended, filterByCategory is alive
