+++ src/components/ContactForm/ContactForm.module.css
@@ -1,0 +1,13 @@
+.container {
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+  padding-left: 30px;
+}
+
+.background {
+  background-size: 100%;
+  margin: 10px;
+  border-radius: 2rem;
+  padding: 6px;
+}
