+++ src/redux/action-type.js
@@ -20,4 +20,6 @@ export const SET_TOTAL_PAGES = "SET_TOTAL_PAGES";
 
 export const  SET_PAGE = "SET_PAGE";
 
+export const  SET_CATEGORY = "SET_CATEGORY";
+
 export const BASE_URL = "https://back-server-pinturas-app.onrender.com/";
