+++ src/redux/actions/filters/setCategory.js
@@ -1,0 +1,6 @@
+import { SET_CATEGORY } from '../../action-type';
+
+export const setCategory = (category) => ({
+    type: SET_CATEGORY,
+    payload: category,
+});
